Added skipping directories in copying user resources. Expanded dummy display list. Fixed checklist values returning as true. Updated linear track model to develop version. Fixed strips in reco_resources_bundle. Added ResourceStorage.has_resources() method.

+++ application.py
@@ -331,7 +331,8 @@ class PADAMOReco(QMainWindow):
             print("COMMONS TGT DIR", tgtdir)
             for file in os.listdir(STOCK_COMMONS_SRCDIR):
                 srcfile = os.path.join(STOCK_COMMONS_SRCDIR, file)
-                shutil.copy(srcfile, tgtdir)
+                if os.path.isfile(srcfile):
+                    shutil.copy(srcfile, tgtdir)
 
     def _run_parallel(self):
         try:
